requirements.txt
streamlit
pandas
fpdf2
extract-msg
openai
python-dotenv